+++ server/routes/adjudications/index.ts
@@ -1,62 +1,66 @@
 import { Request, Response, Router } from 'express'
 
+import { asyncHandler } from '../../middleware/asyncHandler'
+import featureFlagMiddleware from '../../middleware/featureFlag/featureFlag'
+
 import type { Services } from '../../services'
+
 import { formatReportedAdjudication } from '../../utils/adjudications/formatReportedAdjudication'
-import { asyncHandler } from '../../utils/asyncHandler'
 import { getEstablishmentLinksData } from '../../utils/utils'
 
 export default function routes(services: Services): Router {
   const router = Router()
 
-  const renderAdjudicationsPage = async (req: Request, res: Response) => {
-    const { prisonerContentHubURL } = await getEstablishmentLinksData(res.locals.user.idToken.establishment.agency_id)
-    const reportedAdjudications = await services.prisonerProfileService.getReportedAdjudicationsFor(res.locals.user)
-    const displayPagination: boolean = reportedAdjudications.totalPages > 1
+  router.get(
+    '/',
+    featureFlagMiddleware('adjudications'),
+    asyncHandler(async (req: Request, res: Response) => {
+      const { user } = res.locals
 
-    res.render('pages/adjudications', {
-      givenName: res.locals.user.idToken.given_name,
-      title: 'Adjudications',
-      config: getConfig(),
-      data: {
-        reportedAdjudications,
-        displayPagination,
-        adjudicationsReadMoreURL: `${prisonerContentHubURL}/content/4193`,
-      },
-      errors: req.flash('errors'),
-      message: req.flash('message'),
-    })
-  }
+      const { prisonerContentHubURL } = getEstablishmentLinksData(user.idToken.establishment.agency_id) || {}
+      const reportedAdjudications = await services.prisonerProfileService.getReportedAdjudicationsFor(
+        user.idToken.booking.id,
+        user.idToken.establishment.agency_id,
+      )
+      const displayPagination = reportedAdjudications.totalPages > 1
 
-  const renderAdjudicationPage = async (req: Request, res: Response) => {
-    const { prisonerContentHubURL } = await getEstablishmentLinksData(res.locals.user.idToken.establishment.agency_id)
-    const { reportedAdjudication } = await services.prisonerProfileService.getReportedAdjudication(
-      req.params.chargeNumber,
-      res.locals.user.idToken.establishment.agency_id,
-    )
-    const formattedAdjudication = await formatReportedAdjudication(reportedAdjudication, services)
+      res.render('pages/adjudications', {
+        givenName: user.idToken.given_name,
+        title: 'Adjudications',
+        data: {
+          reportedAdjudications,
+          displayPagination,
+          readMoreUrl: `${prisonerContentHubURL}/content/4193`,
+        },
+        errors: req.flash('errors'),
+        message: req.flash('message'),
+      })
+    }),
+  )
 
-    res.render('pages/adjudication', {
-      givenName: res.locals.user.idToken.given_name,
-      title: `View details of ${reportedAdjudication.chargeNumber}`,
-      config: getConfig(),
-      data: {
-        adjudication: formattedAdjudication,
-        adjudicationsReadMoreURL: `${prisonerContentHubURL}/content/4193`,
-      },
-    })
-  }
+  router.get(
+    '/:chargeNumber',
+    featureFlagMiddleware('adjudications'),
+    asyncHandler(async (req: Request, res: Response) => {
+      const { user } = res.locals
 
-  router.get('/', asyncHandler(renderAdjudicationsPage))
-  router.get('/:chargeNumber', asyncHandler(renderAdjudicationPage))
+      const { prisonerContentHubURL } = getEstablishmentLinksData(user.idToken.establishment.agency_id) || {}
+      const { reportedAdjudication } = await services.prisonerProfileService.getReportedAdjudication(
+        req.params.chargeNumber,
+        user.idToken.establishment.agency_id,
+      )
+      const formattedAdjudication = await formatReportedAdjudication(reportedAdjudication, services)
+
+      res.render('pages/adjudication', {
+        givenName: user.idToken.given_name,
+        title: `View details of ${reportedAdjudication.chargeNumber}`,
+        data: {
+          adjudication: formattedAdjudication,
+          readMoreUrl: `${prisonerContentHubURL}/content/4193`,
+        },
+      })
+    }),
+  )
 
   return router
-}
-
-function getConfig() {
-  return {
-    content: false,
-    header: false,
-    postscript: true,
-    detailsType: 'small',
-  }
 }
