+++ server/views/pages/adjudications.njk
@@ -23,7 +23,7 @@
   {% else %}  
       <div class="govuk-!-margin-bottom-7">
         <p class="govuk-body govuk-!-font-size-24 govuk-!-margin-bottom-2">Up to 50 adjudications are shown here, in date order, with the newest first.</p>
-        <a href='{{ data.adjudicationsReadMoreURL }}'  rel="noreferrer noopener" target="_blank" class="govuk-link govuk-!-font-size-24">Read more about adjudications</a>
+        <a href='{{ data.readMoreUrl }}'  rel="noreferrer noopener" target="_blank" class="govuk-link govuk-!-font-size-24">Read more about adjudications</a>
       </div>  
 
       <table class="govuk-table">
@@ -34,6 +34,7 @@
             <th scope="col" class="govuk-table__header"></th>
           </tr>
         </thead>
+
         <tbody class="govuk-table__body">
           {% for adjudication in data.reportedAdjudications.content %}
           <tr class="govuk-table__row">
